app/api/handlers/analysis.py
from typing import Dict, Optional
from uuid import UUID
from datetime import datetime
from uuid import uuid4

from app.api.schemas.analysis import AnalysisRequest, AnalysisResponse, AnalysisList

class AnalysisHandler:
    """Handler para processamento de análises"""
    
    def __init__(self):
        self._analyses = {}  # Simula DB por enquanto
    
    async def create_analysis(self, request: AnalysisRequest) -> AnalysisResponse:
        """Cria uma nova análise"""
        analysis_id = str(uuid4())
        analysis = AnalysisResponse(
            id=analysis_id,
            status="pending",
            content=request.content,
            created_at=datetime.utcnow(),
            updated_at=datetime.utcnow()
        )
        self._analyses[analysis_id] = analysis
        return analysis
    
    async def get_analysis(self, analysis_id: UUID) -> Optional[AnalysisResponse]:
        """Recupera uma análise específica"""
        return self._analyses.get(str(analysis_id))
    
    async def list_analyses(self, page: int, size: int) -> AnalysisList:
        """Lista análises paginadas"""
        items = list(self._analyses.values())
        start = (page - 1) * size
        end = start + size
        
        return AnalysisList(
            items=items[start:end],
            total=len(items),
            page=page,
            size=size
        )
